Update openapi.yaml
Changing the model for page_access_users to correctly reflect an array
of strings coming back for `page_access_group_ids`

I need this for a change to the terraform provider to actually handle
page access users and page access groups.  Perhaps you'll find it useful
too.

I didn't update to the latest swagger.json/openapi.yaml because there
appears to be some redaction at the very least and the upstream
swagger.json is still wrong about the output from this api call.